requirements.txt
tensorflow>=2.15.0
nltk>=3.8.1
PyMuPDF>=1.23.0
numpy>=1.24.0
PyYAML>=6.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.30.0
pytest>=7.4.0
openai>=1.0.0
google-generativeai>=0.8.0
jupyter>=1.0.0
ipykernel>=6.29.0
